+++ Emulator/FileSystems/FSDataBlock.cpp
@@ -102,7 +102,7 @@ OFSDataBlock::writeData(FILE *file, isize size)
 isize
 OFSDataBlock::addData(const u8 *buffer, isize size)
 {
-    isize count = MIN(bsize() - headerSize(), size);
+    u32 count = MIN(bsize() - headerSize(), (u32)size);
 
     memcpy(data + headerSize(), buffer, count);
     setDataBytesInBlock(count);
